// workspace/projects/public-portal/src/app/modules/shared/models/EnumLookuptypes.ts
export enum EnumLookuptypes {
  ApplicantType = "applicantType",
  EducationLevel = "EducationLevel",
  AssetSizeLookup = "AssetSizeLookup",
  AssetTypeLookup = "AssetTypeLookup",
  AwjuhElSarfLookup = "AwjuhElSarfLookup",



  CityLookup = "CityLookup",


  waqfAssetType = "waqfAssetType",


  RegionLookup = "RegionLookup",
  OwnerType = "OwnerType",
  SeerType = "SeerType",
  SifaEtibariType = "SifaEtibariType",
  WaqfTypeLookup = "WaqfTypeLookup",
  IdTypeLookup = "IdTypeLookup",
  AssetSubType = "AssetSubType",
  IssuanceCourtLookup = "IssuanceCourtLookup",
  NationalityLookup = "NationalityLookup",
  BankLookup = "BankLookup",
  ExperienceYear = "ExperienceYear",
  BlackListReason = "BlackListReason",



  EntityType = "EntityType",

  WaqfAssetType = "WaqfAssetType",
  OwnershipTransferReasons = "OwnershipTransferReasonslookup"

}